requirements.txt
APScheduler==3.6.3
argon2-cffi==21.1.0
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
constantly==15.1.0
construct==2.10.54
cryptography==35.0.0
dnspython==2.1.0
docopt==0.6.2
emoji==1.6.1
future==0.18.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
pycparser==2.20
pycryptodomex==3.11.0
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
SecretStorage==3.3.1
six==1.16.0
srv==1.1.0
tornado==6.1
Twisted==21.7.0
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
yarg==0.1.9
zipp==3.6.0
zope.interface==5.4.0
